Start adding some RPCs (#121)
* Start adding some RPCs and update substrate

* Add the AccountStore instead of commenting out assert in substrate

+++ pallets/cash/src/types.rs
@@ -24,8 +24,8 @@ pub type Decimals = u8;
 /// Type for a nonce.
 pub type Nonce = u32;
 
-/// Type for representing time as milliseconds since unix epoch.
-pub type Timestamp = u128; // XXX u64?
+/// Type for representing time since current Unix epoch in milliseconds.
+pub type Timestamp = u64;
 
 /// Type of the largest possible signed integer.
 pub type Int = i128;
